Fix building as dynamic lib with clang
Let's not make things complicated. As of now, qtwebapp only supports
building as dynamic library, even on Windows. Sadly clang was forgotten,
so instead of having a lot of code trying to get import/export right,
build on top of Qt's macros.

QTWEBAPP_PRIVATE is not used in the code, so remove it.